The method itself takes 3 seconds approx. The waitForElements methods are clearly what takes the most time here.
```
16:51:26.551 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated START
16:51:26.589 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated goto(baseUrl) DONE in 0.04s
16:51:27.911 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated waitForElementInWorker(defaultSelector) DONE in 1.36s
16:51:27.991 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated runInWorker(click, defaultSelector) DONE in 1.44s
16:51:29.552 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated Promise.race([this.waitForElementInWorker(loginLinkSelector),this.waitForElementInWorker(logoutLinkSelector)]) DONE in 2.99s
16:51:29.626 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated this.runInWorker(checkAuthenticated) DONE in 3.08s
16:51:29.629 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated END in 3.08s
```

```
16:54:04.252 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated START
16:54:04.311 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated goto(baseUrl) DONE in 0.06s
16:54:05.625 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated waitForElementInWorker(defaultSelector) DONE in 1.37s
16:54:05.715 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated runInWorker(click, defaultSelector) DONE in 1.46s
16:54:07.199 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated Promise.race([this.waitForElementInWorker(loginLinkSelector),this.waitForElementInWorker(logoutLinkSelector)]) DONE in 2.94s
16:54:07.276 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated this.runInWorker(checkAuthenticated) DONE in 3.02s
16:54:07.279 minilog.js:17 Konnector template: 🛡️ ensureAuthenticated END in 3.02s
```